Elastic band working
Added the elastic band path optimization. Unfortunately, it is still possible for the RRT to find a path that is not traversable by a robot. I think adding a radius to the collision check will fix this issue but will add extra computation time as all paths will become more narrow.